[No JIRA] Improve UI tests
The UI tests were running with animation at the default speeds which
caused them to take longer to run. Because we don't care about
animations or their durations in UI tests we can disable animations for
UI tests and increase the speed of those animations that aren't disabled
by this.
