+++ Example/Backpack NativeUITests/ChipUITest.swift
@@ -18,11 +18,7 @@
 
 import XCTest
 
-class ChipUITest: XCTestCase {
-    lazy var app: XCUIApplication = {
-        return XCUIApplication()
-    }()
-
+class ChipUITest: BackpackUITestCase {
     lazy var afghanistanChip: XCUIElement = {
         return app.buttons["Afghanistan"]
     }()
@@ -34,28 +30,6 @@ class ChipUITest: XCTestCase {
     lazy var ethiopiaChip: XCUIElement = {
         return app.buttons["Ethiopia"]
     }()
-
-    override func setUp() {
-        // Put setup code here. This method is called before the invocation of each test method in the class.
-
-        // In UI tests it is usually best to stop immediately when a failure occurs.
-        continueAfterFailure = false
-
-        // UI tests must launch the application that they test.
-        // Doing this in setup will make sure it happens for each test method.
-        XCUIApplication().launch()
-
-        // In UI tests it’s important to set the initial state -
-        //such as interface orientation - required for your tests before they run.
-        // The setUp method is a good place to do this.
-    }
-
-    func saveScreenshot(named name: String) {
-        let screenshot = XCUIScreen.main.screenshot()
-        let attachment = XCTAttachment(screenshot: screenshot)
-        attachment.name = name
-        add(attachment)
-    }
 
     func navigateAndShow() {
         app.tables.staticTexts["Chips"].tap()
