+++ Example/Backpack NativeUITests/DialogUITest.swift
@@ -18,11 +18,7 @@
 
 import XCTest
 
-class DialogUITest: XCTestCase {
-    lazy var app: XCUIApplication = {
-        return XCUIApplication()
-    }()
-
+class DialogUITest: BackpackUITestCase {
     lazy var dialogScrimView: XCUIElement = {
         return app.otherElements["dialogScrimView"]
     }()
@@ -30,28 +26,6 @@ class DialogUITest: XCTestCase {
     lazy var dialogView: XCUIElement = {
         return app.otherElements["dialogView"]
     }()
-
-    override func setUp() {
-        // Put setup code here. This method is called before the invocation of each test method in the class.
-
-        // In UI tests it is usually best to stop immediately when a failure occurs.
-        continueAfterFailure = false
-
-        // UI tests must launch the application that they test.
-        // Doing this in setup will make sure it happens for each test method.
-        XCUIApplication().launch()
-
-        // In UI tests it’s important to set the initial state -
-        //such as interface orientation - required for your tests before they run.
-        // The setUp method is a good place to do this.
-    }
-
-    func saveScreenshot(named name: String) {
-        let screenshot = XCUIScreen.main.screenshot()
-        let attachment = XCTAttachment(screenshot: screenshot)
-        attachment.name = name
-        add(attachment)
-    }
 
     func navigateAndShow(story: String) {
         app.tables.staticTexts["Dialogs"].tap()
