+++ Example/Backpack NativeUITests/HorizontalNavigationUITest.swift
@@ -18,11 +18,7 @@
 
 import XCTest
 
-class HorizontalNavigationUITest: XCTestCase {
-    lazy var app: XCUIApplication = {
-        return XCUIApplication()
-    }()
-
+class HorizontalNavigationUITest: BackpackUITestCase {
     lazy var flightsOption: XCUIElement = {
         return app.buttons["Flights"]
     }()
@@ -30,28 +26,6 @@ class HorizontalNavigationUITest: XCTestCase {
     lazy var hotelsOption: XCUIElement = {
         return app.buttons["Hotels"]
     }()
-
-    override func setUp() {
-        // Put setup code here. This method is called before the invocation of each test method in the class.
-
-        // In UI tests it is usually best to stop immediately when a failure occurs.
-        continueAfterFailure = false
-
-        // UI tests must launch the application that they test.
-        // Doing this in setup will make sure it happens for each test method.
-        XCUIApplication().launch()
-
-        // In UI tests it’s important to set the initial state -
-        //such as interface orientation - required for your tests before they run.
-        // The setUp method is a good place to do this.
-    }
-
-    func saveScreenshot(named name: String) {
-        let screenshot = XCUIScreen.main.screenshot()
-        let attachment = XCTAttachment(screenshot: screenshot)
-        attachment.name = name
-        add(attachment)
-    }
 
     func navigateAndShow() {
         app.tables.staticTexts["Horizontal navigation"].tap()
